Switch to using flake8 for linting.
Flake8 wraps both pylint and pep8 in one call, simplifying the makefile
user experience down to just `make lint`.  It also automatically does
the scanning, removing the need for the find call and pipe to xargs.  It
also allows using #noqa to mark things that don't need to be checked, so
linting the files but ignoring unused imports in __init__ is supported.

- Adds lint target using flake8
- Adds check for presence of flake8 program
- Adds noqa lines in init files