Add documentation to OTIO (#35)
* Initial commit of documentation system.

* Add targets to makefile.

* Add overview and link to otio wiki.

For now we're putting most of the 'real' otio documentation on the
github wiki, so I added a link to that in the overview for the landing
page of the documentation.  For now, the sphinx documentation will be
more for developer API reference.

* Flesh out documentation in the library.

Adds docstrings on many of the methods, modules, functions and classes.
Also adds a docstring to deprecated fields marking them deprecated.

* Alphabetize the module index correctly.

the modindex_common_prefix option lets you ignore a module prefix; added
opentimelineio to this list to take advantage of this feature.

* Include links to source in docs.

* From_filepath returns the adapter object, not name.

Previously it was returning the name, which didn't make sense.  This
change fixes it to return the object and also addresses a bug in
otioconvert that wasn't using the result of the call.

* Address review notes from @joshm.

Documentation related code review notes.