Write media_references from shot_detect.
The examples/shot_detect script thas some useful stuff in it that detects frame ranges and rates.

This adds an argument that will write out the file put on the input as a single clip, with a valid media reference to the entire thing.

Useful for generating test OTIO data from existing files on disk.