Initial port of AAF adapter from using pyaaf to pyaaf2 (#396)
Switched from using pyaaf, to pyaaf2, which is a pure python reimplementation of the AAF library. 

This switch was made in order to make building & installing this easier (many people have trouble getting the AAF libraries to work) and to provide more flexibility into the future.

There are some subtle differences in the AAF metadata that the adapter places into the resulting OTIO, but this seems harmless.

Performance is not as good as it was before, but we hope that we will be better able to profile and optimize this after this switch is complete.

The only outstanding issue that could prevent this from being part of an upcoming OTIO release, is that the pyaaf2 install is done via a git clone from the master branch of pyaaf2. Once the next release of pyaaf2 comes out, we can switch to a regular pip install via pypi (in a future PR to OTIO).

Thanks @freesonluxo and @shahbazk8194 for making this change!