feat(ale_adapter): Add VIDEO_FORMAT to header when writing
Adds a video_format kwarg to the adapter signature, allowing you to set
the format on output. The fallback is to preserve the existing header
column, if set. Finally it searches each clips ALE metadata for an
"Image Size" setting to infer a project format. If none of that is found
it defaults to the adapters DEFAULT_VIDEO_FORMAT (1080)