Pass track name to mobslot (#500)
This value is shown as track name in the Media Composer.

pyaaf2 gave default name as `"Track-%03d" % slot_id`. Here we just pass whatever name a track has to mobslot.

If the name is None, the Media Composer does not show the name (after `/`), but still labels the track as V1, V2 for video tracks and A1, A2 for audio tracks.