Fix source clip start offset when writing AAFs (#506)
* Fix missing source clip start offset

Without this, the resulting AAF always use the full available media of a clip even the start_time in source_range has different value than that of availabe_range.

Fixed private method `_target_url_fixup`.
It didn’t really use the `timeline` passed in. It just happened to pick up the`otio_timeline` when the method was called due to Python scoping.