xges: Fix tracks ordering
The tracks should be loaded in reverse order for the XGES format.

Also in GES, layers (tracks in otio terms) do not have a track type
(tracks are global to the timeline which formalizes the outputed
types), thus we should try to keep audio and video tracks in the same
order during deserialization.