xges: re-implement GstStructure as an otio schema
GstStructure was re-implemented as an otio schema and was improved to
closer emulate the Gstreamer functions 'gst_structure_to_string' and
'gst_structure_from_string'.

As an otio schema, it can now be written to an otio_json file, which
will be useful for making the properties and metadatas xges attributes
more transparent after transforming.

The methods 'serialize_string' and 'deserialize_string' were added to
correctly parse and print string values that contain non-basic-ascii.
In particular, utf8 text can be used for, say, the .name of otio Schema
objects, be set as a property of an xges element, and then be correctly
read by the GES library.

Also:
+ added type checking and other safety measures.
+ made changes to where GstStructure is called.
+ used GstStructure in XgesTrack.