xges: reimplement clip and transition timings
The previous implementation did not trim/delay clips to take into
account that xges and otio handle transitions quite differently. In xges
clips will literally overlap in their durations, but in otio they will
not, so the translation from xges "start", "duration" and "inpoint"
values to otio's "source_range" is non-trivial.

Also added a test to ensure the timings are correctly translated.