xges: bundle project and timeline in otio stack
Placed the xges project and timeline data under a single otio stack,
rather than a serializable collection and an otio timeline, as before.
This will allow us to later support sub-projects, which will be
represented by sub-stacks. This is necessary because sub-timelines are
not supported by otio (they are not composables).

A side benefit is that read_from_string() will now actually return an
otio timeline, rather than a serializable collection.

Also bundled the xges track data inside the otio stack.