xges: merge otio tracks if equal modulo kind
If we have two otio tracks that are the same except their kinds are
different (Audio and Video), then we want to be able to merge the two
into a single track, such that all corresponding xges clips will simply
have track-types=6. As such, we can avoid creating two separate layers
with identical clips (except for the track-types). In particular,
converting an xges layer that only contains clips with track-types=6
will convert to otio and back to a single layer. Note, this won't be
true if a layer contains clips with different track-types because they
will usually produce two tracks that differ by more than track kind.

In order to do this successfully and cleanly, the otio composable names
are no longer required to be unique per stack. Note, there does not seem
to be any requirement that the otio names remain unique.