Ensure that dictionary comparison is symmetrical
The any dictionary equality function was walking over the left
dictionary and incrementing an iterator.  To make sure that the
dictionaries are equivalent, checking to make sure that the end of BOTH
dictionaries has been reached at the end of the function instead of just
the left dictionary should correct the symmetry of is_equivalent_to.