xges: handle unknown types in GstStructure
Show a warning if GstStructure finds an unknown data type (including
ranges, lists and arrays when their elements contain (un)known types).
Before, we would simply raise an exception. But it is quite reasonable
that some metadata or GstCaps entry contains an unknown data type, so
instead we simply store the the data in its serialized form as a string
so that it can be simply copied when writing the GstStructure to a
string.

In addition, we are now converting all given types to their default
alias. This way we can easily recognise that two GstStructures are
equivalent, even if the user gave different aliases.