xges: add GstCaps and GstCapsFeatures schemas
GstCaps is a schema that holds an array of GstStructures, with each
being associated with some GstCapsFeatures.

The XgesTrack schema now uses GstCaps for the caps attribute.
In addition, the _get_from_caps method in XGES now uses the GstCaps
class to search for the given field.

GstCaps are also used in the GES library for serializing/deserializing
markers, so their inclusion will make it easier to eventually extract
markers from the xges element metadatas attribute.