Adding support for negative time strings when converting a negative RationalTime object to a string.
Before this change, doing

negativeRationalTime = otio.opentime.RationalTime(-24,24)
timeString = otio.opentime.to_time_string(negativeRationalTime)

timeString would be "-1:-1:-10", which is not a useful time string.

This change will create a 'negative' time string, i.e. generate the
same string value as a positive RationalTime value would do but
with a '-' sign in front of the string:

timeString is now: '-00:00:01.0'

This is actually a valid string format for ffmpeg and useful if you
for example want to add silence to an audio file at the beginning.