xges: fix handling of sub-stacks
Previously, the method would assume that the asset-id stored in the otio
metadata by the xges adapter was sufficient to determine whether the
sub-project found in the otio stack was equivalent. However, the otio
object may be edited in between, which can lead to different
sub-projects. Although, not always.
Also, there was a bug in the previous code that meant a None asset id
was returned by _serialize_stack_to_ressource if the asset id already
existed.
Instead, we now explicitly test whether the corresponding sub-projects are
equivalent, and share the asset id only if they are.

Added a test for these cases.