xges: Effects and Markers Support (#609)
* xges: add methods for children-properties

<effect>s and GESTransitionClip <clip>s both have children-properties.

* xges: convert some methods to static/classmethods

For methods that didn't use the class instance itself and only perform
a supplementary role.

* xges: _add_properties_and_metadatas_to_element()

Adding this method reduces repeated code.

* xges: increase ges version to 0.6

xges files that include sub-projects are at least version 0.6. We'll
assume this highest version.

* xges: fix ensuring clip name is unique

Previously, names found in the otio metadata properties were not checked
to be unique, leaving a window open for two clips to have the same name.
Also, we don't need the project name to be unique.

* xges: put xges clip properties under "clip" key

Moved the clip properties and metadatas to a sub-dictionary within the
otio metadata so that an otio_stack can also store the properties and
metadatas for corresponding xges clip.

* xges: add tests for properties and metadatas

Added some tests to ensure properties and metadatas are being correctly
set on xges elements.
Also added some assertions to ensure all clips/assets/layers elements
contain certain attributes.

* xges: add support for effects

Added basic support for effects found under xges clips/under otio items.
Moved setting the name and metadata of otio_composables to
_create_otio_composables_from_layer_clips() to reduce repeated code.
Also added some warnings if unsupported track_types are found.

* xges: rename _reference_from_id

Rename to _otio_reference_from_id to follow the naming convention of
other methods.

* xges: support layer properties and metadatas

Store layer properties and metadatas in otio tracks.
Also namespace the otio tracks (as otio_track, etc) and otio stacks in
XGES, to follow the naming convention, and help distinguish them from
the xges tracks (which are quite different) and the track_type(s).

* xges: tidy GstStructure schema

In particular, the initialization method is now split in two to make it
more explicit. The default initialization accepts the name and fields
arguments as they would be stored internally. For string values, the
new_from_str class method is used for deserializing.

Also:
+ Added the exception DeserializeError.
+ Moved conversions of unicode types in python2 to the function
  unicode_to_str. This should make it easier to remove this step once
  python2 has ended.
+ Added new methods, and use them in the adapters

* xges: add GstCaps and GstCapsFeatures schemas

GstCaps is a schema that holds an array of GstStructures, with each
being associated with some GstCapsFeatures.

The XgesTrack schema now uses GstCaps for the caps attribute.
In addition, the _get_from_caps method in XGES now uses the GstCaps
class to search for the given field.

GstCaps are also used in the GES library for serializing/deserializing
markers, so their inclusion will make it easier to eventually extract
markers from the xges element metadatas attribute.

* xges: support serialized structures and caps

Add support for finding serialized structures or caps in a GstStructure
string. This allows for nested structures/caps.

* xges: support markers

Added the GESMarker and GESMarkerList schemas and used them to support
converting xges markers into otio markers, and back.

* xges: convert track effects to GESEffectClips

Convert a track effect to a GESEffectClip that covers all the other
clips in the layer.

* xges: test for layer-priority and unique id

Test that clips have the correct layer-priority and have unique ids.

* xges: fix test documentation

* xges: remove unused objects

The value and method were unused.

* xges: mark private methods

Prepended a '_' to some private methods

* xges: add some internal documentation

This should make it easier to maintain, and for new contributors.

* xges: test that unsupported track types

Make sure that unsupported track types found in xges do not cause
problems for the other track types.

* xges: fix handling of sub-stacks

Previously, the method would assume that the asset-id stored in the otio
metadata by the xges adapter was sufficient to determine whether the
sub-project found in the otio stack was equivalent. However, the otio
object may be edited in between, which can lead to different
sub-projects. Although, not always.
Also, there was a bug in the previous code that meant a None asset id
was returned by _serialize_stack_to_ressource if the asset id already
existed.
Instead, we now explicitly test whether the corresponding sub-projects are
equivalent, and share the asset id only if they are.

Added a test for these cases.