Update AAF reader to discover metadata (UserComments) in more cases, only do 1-pass transcribe.
This change to the AAF reader is mainly motivated by discovering relevant metadata from clips for some (special) case scenarios which have not been covered by the current code.

When traversing the AAF object-tree, at the point we reach a SourceClip, we try to discover the masterMob to then extract the metadata (UserComments) from the MasterMob.

The way we discover the MasterMob is that either the referenced mob in the SourceClip is the masterMob or we walk pack the parent-chain to find the masterMob.

However, there is an additional case, where the SourceClip can reference a CompositionMob which then references the masterMob. This was first discovered in one of our AAFs using subclips.

This aaf-reader-update now includes this CompositionMob discovery, so when subclips are used, the expected UserComments metadata is discovered properly.

In addition to this discovery there are two more scenarios:

The SourceClip can reference a CompositionMob, but instead of holding a masterMob, the CompositionMob itself told the UserComments. I believe this is a way AVID realizes sharing of the same metadata between SourceClips.

The SourceClip can reference a CompositionMob with UserComments AND hold a MasterMob. In this case the MasterMob seems to represent the valid UserComments, so in this scenario, we use the UserComments of the MasterMob.

I added three test-cases for the scenarios described above:

test_aaf_subclip_metadata()
test_aaf_composition_metadata()
test_aaf_composition_metadata_mastermob()

In addition I added another debug flag, transcribe_debug. If enabled, it will output the recursive traversal of
_transcribe(). This is helpful in order to see and understand the traversal-process.

Finally I removed the first transcribe-pass and instead of building up a cache of MasterMobs, transcribe the MasterMob in the second pass 'on-demand'. Since the first pass used to transcribe any AAF object
(and then basically threw away the results) this improves the time it takes to parse the AAF by a fairly large number (around 30-40%)