AAF adapter: added `clip_timecode_offset` argument
Added `clip_timecode_offset` as new optional argument.
By default, the behaviour is the same as before.
However, if the parameter is set to `False`, the global
timecode offset will not be added to the clip source range. This allows for easy access to the actual
source range used in the clip.