Simplifie wheel target
Replacing wheel build command from:
`python setup.py bdist_wheel` with:
`pip wheel . --no-deps`
This will have several advantages.

Use pyproject for dependencies. This makes it much cleaner to build
wheels.

We no longer need to check for the availability of the wheel package.
And the build will be out of source, and we don't need to cleanup
intermediate folders.

The `clean` target will now remove *.whl files in the root of the
project. This is where `make wheel` will output the wheels to with the
new command.